rag_chatbot/requirements.txt
llama-index
fastapi
uvicorn
langgraph
langchain-pinecone
langchain-openai
nest-asyncio
pypdf
llama-parse

pinecone
pymongo

unstructured[pdf]
python-dotenv
pydantic
httpx
openai
tiktoken
